Refactored code to remove the term "batch_wrapper", since this no longer runs in batch.  Generically, it processes reqeusts from the web application.
