+++ Dockerfile
@@ -24,5 +24,3 @@ RUN pip install pyyaml
 RUN pip install certifi
 RUN pip install urllib3
 RUN easy_install-3.7 fsoi-0.1-py3.7.egg
-
-CMD ["batch_wrapper"]
